Initial code commit for git
This is essentially the same as the 1.2 release (no code changes), but license text is added in all .hx files and extra information is provided in the readme file (which now is in text format). Also a license.txt is added with only the license code.

The HTML documentation is not included in the repository because it can be generated using SlashDOC from the comments (so it is not "source code"). It can still be read online at Flash MOD Player's page and of course in the source code comments.