Removed unneeded breaks and reformatted the code
I reformatted the code a bit to be more consistent with my newer coding style (the current code was a mix of the old and new). Also i removed two unneeded 'break's from the command switch for commands 0x01 "slide up" and 0x02 "slide down" (in haXe 'switch' do not need break).