Modified .gitignore to ignore MOD files
I regularly use MOD files for testing and keep them in the source directory. However, as one can imagine, i don't want them to be commited to the repository, so i add this entry in .gitignore to ignore these files. You really don't want to hear what i'm listening to while trying to debug the player - and also i don't want you to hear MODs that play incorrectly :-).

Beyond this small addition in the .gitignore, i also commit Test.hx because git keeps nagging me about it being modified and such. I didn't really modified it. Ok, i did. But only a little. I only changed the filename of the MOD to load. But is this really a good reason to nag me git? Is it?

I wonder...