Fixed a bug in the 0x0B 'jump' command
Some mods use the 0x0B command to jump to the next pattern before the current pattern is finished and the next pattern contains speed modifying commands. Previously the code updated 'corder' and 'crow' but forgot to update 'cpat' and 'cpattern' (sidenote: shouldn't these two be renamed?). The command sets up the 'nextcheckrow' flag which makes the row check (for commands, updates, etc) to be enforced in the next turn, but it the check wasn't done in the following (target) row because the following row was the first row of the current pattern instead of the target pattern. This led to two side effects: first, because the pointers were not updated and the 'crow' was set to 0, the pattern was played twice (but not forever because the corder was updated). Secondly, since 'nextcheckrow' wasn't effective the program *jumped* the call to the speed modifying commands - the chain result was playback that was timed incorrectly. In some cases this wasn't very noticeable (well, except to the original artists :-P) but in some other cases, the playback was horrible.

So now this bug is fixed with a nice fix: just moving two lines outside an 'if' check :-P. It seems to work and the awesome MOD 'Danny elfmania' by mobi sounds great. See by yourself at http://modarchive.org/module.php?36042 :-)