Added support for xxCH MODs and added a null check
Although the code supported them, there wasn't a test for MODs which had a "xxCH" signature (these are mods of 'any' number of channels from 1 to 99 and xx is the channel count number). So i added this test which makes MODs which use unusual sets of channels (7, 12, 15, etc) to be supported.

Also a null check in note retriggering is added which checks if the current sample is null or not. However this might not be the best solution given that a note retrigger is supposed to have a sample set.