Fixed some sample-related command bugs
Some commands (E9 and ED to be specific) triggered notes outside of the "normal" time. Usually when a note is triggered it also sets the current volume for the channel. However the implementation of the above commands didn't and so in some cases the volume was not correct. This is fixed in this commit.

Another issue that was fixed was when a sample needs to be "cut out", its reference (cs.csmp) was set to null, which caused sample retriggers (E9) to not work (before a previous commit this raised an exception which effectivelly stopped the presynthesis - later it was just ignored). Now the code simply sets the volume to 0 which essentially has the same effect, but note retriggering works properly.