Fixed some bugs and added name field in samples
A new field in the Sample class is added that contains the sample's name. This will be used in a later update to the API so a player can obtain information about the MOD.

Also some bugs were fixed. One was related to the fixed-point math that is done (i'm not 100% sure that fp math is really required actually, but lets pretend for now that it is - besides if FMP is ported to other platforms at the future, like Android or some ARM-based devices, it might be a good idea to use fixed point instead of floats). The sample counters, lengths, etc were done in 16.16 fixed-point math, but in Flash Int is singned so one point was lost. To avoid that (and other clipping issues), the fixed point variables are now UInt and the fixed point math is done in 18.14, which should be more than enough.

The second bug was in the 0E 'pattern break' command. The command was delayed until the next row before, but now it is handled in the next 'cycle' of the main loop.